LayoutTests/imported/w3c/web-platform-tests/css/css-cascade/scope-supports-expected.txt


FAIL Style rules within @supports are scoped assert_equals: expected "1" but got "auto"
